backend/requirements.txt
# Development
Flask
pydantic
flasgger
pyjwt
flask-cors
Flask-SQLAlchemy
pytz

# Testing
coverage
requests
pytest

# Style
black